requirements.txt
asgiref==3.6.0
Django==4.1.7
django-widget-tweaks==1.4.12
fontawesomefree==6.4.0
mysqlclient==2.1.1
Pillow==9.4.0
sqlparse==0.4.3
